move huge array to heap because it might overflow the stack size
When using the library in the variant with the largest public key
(namely mceliece8192128), the stack allocation amounts
to 1,703,936 bytes. This is 20% of the maximum admissible stack size
of my system. This high stack consumption is a problem on several
systems. Thus, I propose to allocate this huge array on the heap.

Context:
  issue #13
  https://github.com/Colfenor/classic-mceliece-rust/issues/13