feature: split tests in lib.rs into separate files
The updated version exposes the same API and runs the same set of tests,
but the tests have been moved to separate modules.

According to semver-checks, no API breakage was found after running:

  cargo semver-checks --only-explicit-features --baseline-rev main
                      --baseline-features mceliece348864
		      --current-features mceliece348864

Sadly, we cannot move these tests into the tests/ directory,
because the public API does not expose functionality like "fn verify".
As such, we cannot use verify from files in tests/.
Even if we migrate "fn verify" to tests/, there is still AesState.
Since AesState is used in so many place, we can only migrate
the tests to tests/ by code duplication (which I do not want to do).

Furthermore:

• Address TestData with `TestData` after an `use` statement
  instead of `crate::TestData` to make it easier to move TestData around
• a lot of feature guard fixes
• KATKEM tests pass for me with "ulimit -s 12048",
  but if you additionally pass feature="alloc" now,
  the public key will end up on the heap instead of the stack now.
• I ran all tests in all configurations and made sure no performance
  regressions occur.